Corrigindo bugs

+++ mod39/SAPrint/src/main/java/com/samuel/dao/IClienteDao.java
@@ -5,7 +5,7 @@ import java.util.List;
 import com.samuel.dao.generics.IGenericDao;
 import com.samuel.model.Cliente;
 
-public interface IClienteDao extends IGenericDao<Cliente, Integer> {
+public interface IClienteDao extends IGenericDao<Cliente, Long> {
     
 	
 	List<Cliente> filtrarClientes(String query);
