+++ mod39/SAPrint/src/main/java/com/samuel/service/IClienteService.java
@@ -6,9 +6,9 @@ import com.samuel.exception.DaoException;
 import com.samuel.model.Cliente;
 import com.samuel.service.generic.IGenericService;
 
-public interface IClienteService extends IGenericService<Cliente, Integer> {
+public interface IClienteService extends IGenericService<Cliente, Long> {
 	
-	Cliente buscarPorCpf(Integer cpf) throws DaoException;
+	Cliente buscarPorCpf(Long cpf) throws DaoException;
 	List<Cliente> filtrarClientes(String query);
 	
 }
